requirements.txt
# Core dependencies for video processing
opencv-python==4.8.1.78
numpy>=1.24.3
Pillow==10.0.1

# PyTorch (install first, separately if needed)
torch>=2.1.0
torchvision>=0.16.0

# Core ML/AI libraries
scipy>=1.10.0
scikit-image>=0.20.0
matplotlib>=3.6.0
tqdm>=4.64.0
imageio>=2.25.0

# Claude VLM for scene understanding
anthropic>=0.7.8

# Video streaming
ffmpeg-python>=0.2.0

# 3D processing (basic)
open3d>=0.17.0
trimesh>=3.23.5

# Emotion Detection (Multi-Model)
torch>=1.2.0
torchvision>=0.4.0
scikit-image>=0.15.0
face-alignment>=1.1.1
mediapipe>=0.10.0
fer>=22.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
PyYAML>=6.0.0
asyncio
